Rename resotolib.signal -> resotolib.proc
