+++ resotolib/resotolib/config.py
@@ -3,7 +3,7 @@ import threading
 from resotolib.logger import log
 from resotolib.args import ArgumentParser, convert
 from resotolib.core.ca import TLSData
-from resotolib.signal import restart
+from resotolib.proc import restart
 from resotolib.core.model_export import dataclasses_to_resotocore_model, optional_origin
 from resotolib.core import ResotocoreURI
 from resotolib.core.config import (
