+++ resotoshell/resotoshell/__main__.py
@@ -3,7 +3,7 @@ import sys
 from threading import Event, Thread
 from typing import Callable
 
-import resotolib.signal
+import resotolib.proc
 from resotolib.args import ArgumentParser, Namespace
 from resotolib.core import resotocore, add_args as core_add_args, resotocore_is_up
 from resotolib.core.ca import TLSData
@@ -17,8 +17,8 @@ from resotoclient import ResotoClient
 
 
 def main() -> None:
-    resotolib.signal.parent_pid = os.getpid()
-    resotolib.signal.initializer()
+    resotolib.proc.parent_pid = os.getpid()
+    resotolib.proc.initializer()
     setup_logger("resotoshell", json_format=False)
     arg_parser = ArgumentParser(
         description="resoto shell", env_args_prefix="RESOTOSHELL_"
@@ -44,7 +44,7 @@ def main() -> None:
         handle_from_stdin(client)
     else:
         repl(client, args)
-    resotolib.signal.kill_children(resotolib.signal.SIGTERM, ensure_death=True)
+    resotolib.proc.kill_children(resotolib.proc.SIGTERM, ensure_death=True)
     log.debug("Shutdown complete")
     sys.exit(0)
 
@@ -60,7 +60,7 @@ def repl(
 
     def shutdown(event: ResotoEvent) -> None:
         shutdown_event.set()
-        kt = Thread(target=resotolib.signal.delayed_exit, name="shutdown")
+        kt = Thread(target=resotolib.proc.delayed_exit, name="shutdown")
         kt.start()
 
     add_event_listener(EventType.SHUTDOWN, shutdown)
